Major refactoring
Changed Stocks to StockPositions, which belong to a User and are dependent on a Stock. The stock is independent of everything.
Changed all the values to Doubles.
Added an interface for valueItems, which add methods to retrieve the currency and the value.
Added dependency to coroutines.

+++ MoneyManagementBusiness/src/main/kotlin/com/mauquoi/money/business/service/AccountService.kt
@@ -43,7 +43,7 @@ class AccountService @Inject constructor(private val userRepository: UserReposit
         accountAuditRepository.save(new)
     }
 
-    fun updateAccountValue(accountId: Long, amount: Float) {
+    fun updateAccountValue(accountId: Long, amount: Double) {
         val account = getAccount(accountId)
         val new = AccountSnapshot(account = account,
                 amount = amount,
@@ -63,8 +63,8 @@ class AccountService @Inject constructor(private val userRepository: UserReposit
     }
 
 //    todo view how to do this with different currencies
-    fun getTotalAccountValue(userId: Long): Float {
-        return getAccounts(userId).sumByDouble { it.amount.toDouble() }.toFloat()
+    fun getTotalAccountValue(userId: Long): Double {
+        return getAccounts(userId).sumByDouble { it.amount }
     }
 
     fun editAudit(auditId: Long, accountSnapshot: AccountSnapshot) {
