+++ MoneyManagementBusiness/src/main/kotlin/com/mauquoi/money/model/OverviewItem.kt
@@ -2,7 +2,7 @@ package com.mauquoi.money.model
 
 import java.util.*
 
-data class OverviewItem(val mainCurrencyValue: Float,
+data class OverviewItem(val mainCurrencyValue: Double,
                         val mainCurrency: Currency,
-                        val distribution: Map<Currency, Float>
+                        val distribution: Map<Currency, Double>
 )
