+++ MoneyManagementBusiness/src/main/kotlin/com/mauquoi/money/model/audit/Snapshot.kt
@@ -12,7 +12,7 @@ data class AccountSnapshot(@Id
                            @GeneratedValue(strategy = GenerationType.IDENTITY)
                            var id: Long? = null,
                            @ManyToOne @JsonIgnore val account: Account?,
-                           @Column(name = "amount", nullable = false) val amount: Float,
+                           @Column(name = "amount", nullable = false) val amount: Double,
                            @Column(name = "date", nullable = false) val date: LocalDate
 )
 
@@ -22,6 +22,6 @@ data class DepositSnapshot(@Id
                            @GeneratedValue(strategy = GenerationType.IDENTITY)
                            var id: Long? = null,
                            @ManyToOne @JsonIgnore val deposit: Deposit?,
-                           @Column(name = "amount", nullable = false) val amount: Float,
+                           @Column(name = "amount", nullable = false) val amount: Double,
                            @Column(name = "date", nullable = false) val date: LocalDate
 )
