+++ MoneyManagementBusiness/src/main/kotlin/com/mauquoi/money/model/dto/CurrencyLookupDto.kt
@@ -3,4 +3,4 @@ package com.mauquoi.money.model.dto
 import java.time.LocalDate
 import java.util.*
 
-data class CurrencyLookupDto(val base: Currency, val date: LocalDate, val rates: Map<Currency, Float>)
+data class CurrencyLookupDto(val base: Currency, val date: LocalDate, val rates: Map<Currency, Double>)
