+++ MoneyManagementServiceWeb/src/main/kotlin/com/mauquoi/money/controller/AccountController.kt
@@ -48,7 +48,7 @@ class AccountController @Inject constructor(private val accountService: AccountS
 
     @PostMapping(UPDATE_ACCOUNT)
     fun updateAccount(@PathVariable(ACCOUNT_ID) accountId: Long,
-                      @RequestParam(name = "amount", required = true) amount: Float): ResponseEntity<Nothing> {
+                      @RequestParam(name = "amount", required = true) amount: Double): ResponseEntity<Nothing> {
         accountService.updateAccountValue(accountId, amount)
         return ResponseEntity.noContent().build()
     }
